requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
babel==2.17.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==5.1.7
docutils==0.21.2
filelock==3.16.1
Flask==2.2.5
Flask-Babel==2.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.56.0
fpdf2==2.8.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
lime==0.2.0.1
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
networkx==3.4.2
numpy==2.2.4
openai==1.70.0
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
platformdirs==4.3.6
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pyparsing==3.2.3
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.1
reportlab==4.3.1
requests==2.32.3
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
statistics==1.0.3.5
svglib==1.5.1
threadpoolctl==3.6.0
tifffile==2025.3.13
tinycss2==1.4.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2024.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
virtualenv==20.28.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
xgboost==3.0.0
xhtml2pdf==0.2.17
